backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.2
email-validator==2.2.0
pydantic-settings==2.5.2
sqlalchemy==2.0.36
aiosqlite==0.20.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.12
anthropic==0.39.0
chromadb==0.5.20
opentelemetry-sdk==1.28.1
opentelemetry-instrumentation-fastapi==0.49b1
opentelemetry-exporter-otlp-proto-grpc==1.28.1
httpx==0.27.2
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-httpx==0.32.0
